refactor: address history producer interface 생성
