+++ team1-api/src/main/java/goorm/deepdive/team1/api/user/application/UserFacade.java
@@ -28,14 +28,14 @@ import goorm.deepdive.team1.domain.address.application.KakaoApiAddressService;
 import goorm.deepdive.team1.domain.address.domain.Address;
 import goorm.deepdive.team1.domain.addresshistory.application.AddressHistoryCommandService;
 import goorm.deepdive.team1.domain.addresshistory.domain.AddressHistory;
+import goorm.deepdive.team1.domain.addresshistory.infrastructure.AddressHistoryProducer;
 import goorm.deepdive.team1.domain.user.application.UserCommandService;
-import goorm.deepdive.team1.domain.user.application.UserProducer;
 import goorm.deepdive.team1.domain.user.application.UserQueryService;
 import goorm.deepdive.team1.domain.user.domain.User;
 import goorm.deepdive.team1.domain.user.domain.UserCache;
 import goorm.deepdive.team1.domain.user.domain.UserDocument;
 import goorm.deepdive.team1.domain.user.domain.enums.AgeGroups;
-import goorm.deepdive.team1.infra.kafka.producer.AddressHistoryProducer;
+import goorm.deepdive.team1.domain.user.infrastructure.UserProducer;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
