+++ team1-domain/src/main/java/goorm/deepdive/team1/domain/addresshistory/infrastructure/AddressHistoryProducer.java
@@ -1,0 +1,9 @@
+package goorm.deepdive.team1.domain.addresshistory.infrastructure;
+
+import goorm.deepdive.team1.domain.addresshistory.domain.AddressHistory;
+
+public interface AddressHistoryProducer {
+	void sendMessageToCreate(AddressHistory addressHistory);
+
+	void sendMessageToDelete(AddressHistory addressHistory);
+}
