+++ team1-domain/src/main/java/goorm/deepdive/team1/domain/user/infrastructure/UserProducer.java
@@ -1,4 +1,4 @@
-package goorm.deepdive.team1.domain.user.application;
+package goorm.deepdive.team1.domain.user.infrastructure;
 
 import goorm.deepdive.team1.domain.user.domain.User;
