+++ team1-infra/src/main/java/goorm/deepdive/team1/infra/kafka/producer/KafkaAddressHistoryProducer.java
@@ -7,17 +7,19 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import goorm.deepdive.team1.domain.addresshistory.domain.AddressHistory;
+import goorm.deepdive.team1.domain.addresshistory.infrastructure.AddressHistoryProducer;
 import lombok.RequiredArgsConstructor;
 
 @Component
 @RequiredArgsConstructor
-public class AddressHistoryProducer {
+public class KafkaAddressHistoryProducer implements AddressHistoryProducer {
 	private final KafkaTemplate<String, Object> kafkaTemplate;
 	private final ObjectMapper objectMapper;
 
 	private static final String CREATE_ADDRESS_HISTORY = "create-address-history";
 	private static final String DELETE_ADDRESS_HISTORY = "delete-address-history";
 
+	@Override
 	public void sendMessageToCreate(AddressHistory addressHistory) {
 		try {
 			String addressHistoryJson = objectMapper.writeValueAsString(addressHistory);
@@ -27,6 +29,7 @@ public class AddressHistoryProducer {
 		}
 	}
 
+	@Override
 	public void sendMessageToDelete(AddressHistory addressHistory) {
 		try {
 			String addressHistoryJson = objectMapper.writeValueAsString(addressHistory);
